requirements.txt
streamlit==1.31.1
chromadb==0.4.22
PyMuPDF==1.23.26
python-docx==1.1.0
groq==0.4.2
google-generativeai==0.3.2 